+++ src/java/com/cartelle/modelo/Puestos.java
@@ -21,9 +21,9 @@ public class Puestos {
     private String medidasPreventivasExistentes;
     private String observacionesMedidasPreventivas;
     private int idArea;
-    private int luz;
-    private int ruido;
-    private int temp;
+    private float luz;
+    private float ruido;
+    private float temp;
 
     public int getIdPuesto() {
         return idPuesto;
@@ -129,28 +129,30 @@ public class Puestos {
         this.idArea = idArea;
     }
 
-    public int getLuz() {
+    public float getLuz() {
         return luz;
     }
 
-    public void setLuz(int luz) {
+    public void setLuz(float luz) {
         this.luz = luz;
     }
 
-    public int getRuido() {
+    public float getRuido() {
         return ruido;
     }
 
-    public void setRuido(int ruido) {
+    public void setRuido(float ruido) {
         this.ruido = ruido;
     }
 
-    public int getTemp() {
+    public float getTemp() {
         return temp;
     }
 
-    public void setTemp(int temp) {
+    public void setTemp(float temp) {
         this.temp = temp;
     }
+
+   
 
 }
